pyproject.toml
[project]
name = "ai-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.5",
    "fastapi>=0.118.3",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "passlib>=1.7.4",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "slowapi>=0.1.9",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.37.0",
]
